cleanup old files, fix reactive
