+++ syftbox/tui/app.py
@@ -1,14 +1,17 @@
+from pathlib import Path
+
 from textual.app import App
 from textual.widgets import Footer, Header, TabbedContent, TabPane
 
-from syftbox.client.base import SyftBoxContextInterface
-from syftbox.tui.api_widget import APIWidget
-from syftbox.tui.datasites_widget import DatasiteSelector
-from syftbox.tui.home_widget import HomeWidget
-from syftbox.tui.sync_widget import SyncWidget
+from syftbox.lib import Client
+from syftbox.tui.widgets.api_widget import APIWidget
+from syftbox.tui.widgets.datasites_widget import DatasiteSelector
+from syftbox.tui.widgets.home_widget import HomeWidget
+from syftbox.tui.widgets.sync_widget import SyncWidget
 
 
 class SyftBoxTUI(App):
+    CSS_PATH = Path(__file__).parent.parent / "assets" / "tui.tcss"
     BINDINGS = [
         ("h", "switch_tab('Home')", "Home"),
         ("a", "switch_tab('APIs')", "APIs"),
@@ -18,21 +21,13 @@ class SyftBoxTUI(App):
 
     def __init__(
         self,
-        syftbox_context: SyftBoxContextInterface,
-        driver_class=None,
-        css_path=None,
-        watch_css=False,
-        ansi_color=False,
+        syftbox_context: Client,
     ):
+        super().__init__()
         self.syftbox_context = syftbox_context
-        super().__init__(driver_class, css_path, watch_css, ansi_color)
-
-    def action_toggle_dark(self) -> None:
-        self.theme = "textual-dark" if self.theme == "textual-light" else "textual-light"
 
     def action_switch_tab(self, tab: str) -> None:
         self.query_one(TabbedContent).active = tab
-        self.query_one(TabbedContent).active_pane.children[0].focus()
 
     def on_mount(self) -> None:
         self.title = "SyftBox"
@@ -52,3 +47,8 @@ class SyftBoxTUI(App):
             with TabPane("Sync", id="Sync"):
                 yield SyncWidget(self.syftbox_context)
         yield Footer()
+
+
+# config = SyftClientConfig.load()
+# syftbox_context = Client(config)
+# app = SyftBoxTUI(syftbox_context)
